+++ src/main/java/blusunrize/immersiveengineering/common/entities/ChemthrowerShotEntity.java
@@ -23,6 +23,7 @@ import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.material.Fluid;
@@ -179,9 +180,10 @@ public class ChemthrowerShotEntity extends IEProjectileEntity
 		return 0;
 	}
 
+	@Nonnull
 	@Override
 	protected ItemStack getDefaultPickupItem()
 	{
-		return ItemStack.EMPTY;
+		return Items.WATER_BUCKET.getDefaultInstance();
 	}
 }